Reworked to avoid cyclic dependency
