+++ designer/server/src/models/forms/editor-v2/guidance.js
@@ -14,9 +14,9 @@ import {
   SAVE,
   baseModelFields,
   getFormSpecificNavigation,
+  getPageConditionDetails,
   getPageNum
 } from '~/src/models/forms/editor-v2/common.js'
-import { getPageConditionDetails } from '~/src/models/forms/editor-v2/page-conditions.js'
 import { editorv2Path, formOverviewPath } from '~/src/models/links.js'
 
 /**
